Fix failing tests
When adding new modules that are a dependency in the main app, the
modules need to be imported into the test harness. If there are no tests
for the module that import it, then the best place to put the import is
in [spec/spec_index.js]('crossroads.net/spec/spec_index.js').