US6972 Microclient Routing
When the content router can't find the route in the cms, force the
browser to reload with the new route and a query param `resolve=true".
This way the server can determine if there is a microclient with that
route and the query param tells it to show a 404 page if not (don't
fall back to crds-angular or that will create a endless loop).