Children are now getting put into KC Age\group
When this script was written the Kids Club Age\Grade group for each child
was getting inserted into the group role column. Also the scripts was not
adding anything to the Group_ID Column.

Now the script will add the correct data to the group ID to put the
children into thier correct Age\Grade Group for Kids club. Also the will
have the correct role for the group.